Add files via upload
This ZIPFILE is the updated and shortened version of our project. I have added the 3 text files because I was unsure which one to use. I didn't want to just upload the new files and have them be confusing with the old files. Tomorrow we can upload the files after we delete the old ones. That way we can use bits and pieces as needed.